+++ src/events/BeforeCreateOrderEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace QD\commerce\webshipper\events;
+
+use yii\base\Event;
+
+class BeforeCreateOrderEvent extends Event
+{
+	// Properties
+	// =========================================================================
+
+	public $order;
+	public $shippingData;
+}